feat(LinkTo): new CSS setup, no snapshot tests, some package fixes
