+++ package.json
@@ -54,12 +54,15 @@
 		"@storybook/addon-links": "^6.4.9",
 		"@storybook/addons": "^6.4.9",
 		"@storybook/react": "^6.4.9",
+		"@storybook/testing-react": "^1.2.2",
 		"@testing-library/jest-dom": "^5.16.1",
+		"@testing-library/react": "^12.1.2",
 		"@types/react": "^17.0.38",
 		"@types/react-dom": "^17.0.11",
 		"@types/react-test-renderer": "^17.0.1",
 		"@types/styled-components": "^5.1.19",
 		"babel-loader": "^8.2.3",
+		"cssjson": "^2.1.3",
 		"eslint": "^7.32.0",
 		"eslint-config-prettier": "^8.3.0",
 		"eslint-plugin-react": "^7.26.1",
@@ -70,16 +73,12 @@
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2",
 		"react-is": "^17.0.2",
+		"react-test-renderer": "^17.0.2",
 		"rollup-plugin-inline-svg": "^2.0.0",
 		"size-limit": "^7.0.5",
+		"styled-components": "^5.3.3",
 		"tsdx": "^0.14.1",
 		"tslib": "^2.3.1",
 		"typescript": "^4.5.4"
-	},
-	"dependencies": {
-		"@storybook/testing-react": "^1.2.2",
-		"@testing-library/react": "^12.1.2",
-		"react-test-renderer": "^17.0.2",
-		"styled-components": "^5.3.3"
 	}
 }
