+++ src/components/LinkTo/LinkTo.tsx
@@ -3,7 +3,7 @@ import Icon from "../Icon/Icon";
 import { iconTypes } from "../Icon/collection";
 import color from "../../styles/colors";
 import { LinkToProps } from "./types";
-import LinkToStyles from "./LinkTo.styles";
+import { LinkToStyles } from "./LinkTo.styles";
 
 const { LinkStyled, FlexSpanStyled } = LinkToStyles;
 
@@ -18,7 +18,7 @@ const LinkTo: React.FC<LinkToProps> = ({
 		href={`${type === "email" ? "mailto:" : ""}${address}`}
 		target={`${type === "email" ? "_self" : "_blank"}`}
 	>
-		<FlexSpanStyled iconLayout={iconLayout}>
+		<FlexSpanStyled iconLayout={iconLayout} data-testid="test-link-flex">
 			<Icon
 				svg={type === "email" ? iconTypes.mail : iconTypes.link}
 				fill={color.blue}
